image_processor/requirements.txt
requests
google-cloud-storage
google-api-core